starting on this, extracting the set field into its own class
